Type long may be required for large alignments.
git-svn-id: https://pbtech-vc.med.cornell.edu/public/svn/icb/trunk/goby@26632 60cd2ccc-2b05-0410-9566-da273a940bff

+++ src/edu/cornell/med/icb/goby/modes/SequenceVariationStats2Mode.java
@@ -144,7 +144,7 @@ public class SequenceVariationStats2Mode extends AbstractGobyMode {
                 final double totalNumberOfVariationBases = sum(readIndexVariationTallies);
                 final double numberOfAlignmentEntries = iterator.getNumAlignmentEntries();
 
-                final int countReferenceBases = iterator.getReferenceBaseCount();
+                final long countReferenceBases = iterator.getReferenceBaseCount();
                 int maxReadIndex = iterator.getMaxReadIndex();
                 for (int readIndex = 1; readIndex < maxReadIndex + 1; readIndex++) {
                     final int countVariationBasesAtReadIndex = readIndexVariationTallies[readIndex];
